Add script to generate demo GIF from screenshots
This commit introduces a new script, `make-demo-gif.sh`, which automates the creation of a demo GIF using three specified screenshots. The script checks for the existence of the required image files and utilizes either ImageMagick or ffmpeg to generate the GIF. If neither tool is available, it falls back to a Node.js generator. Instructions for installing the necessary tools are provided in case of failure.